weather-station/requirements.txt
flask==0.12.5
Adafruit-DHT==1.4.0
RPi.GPIO==0.7.0
